renamed board-based common file, updated references. added label and description to timer prop
