+++ components/trello/sources/common-polling.js
@@ -9,6 +9,8 @@ module.exports = {
       default: {
         intervalSeconds: 60 * 15,
       },
+      label: "Timer",
+      description: "The frequency in seconds you'd like to poll for new events."
     },
   },
 };
